Added bootstrap image carousel
