+++ app/assets/javascripts/application.js
@@ -16,4 +16,4 @@
 //= require_tree .
 // require jquery3
 //= require popper
-//= require bootstrap-sprockets
+//= require bootstrap
