+++ app/assets/stylesheets/application.scss
@@ -11,13 +11,13 @@
  * It is generally better to create a new file per style scope.
  *
  */
-
 @import "bootstrap";
 @import "shared/variables";
 @import "shared/navbar";
-@import "games/listing";
+@import "games/index";
 @import "games/sell";
 @import "games/edit";
+@import "games/show";
 @import "stripe/charges";
 @import "users/profile";
