+++ app/assets/stylesheets/games/show.scss
@@ -1,0 +1,21 @@
+.slideshow-container {
+  height: 200px;
+  width: 100%;
+  display: flex;
+  align-items: center;
+
+
+  img {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+  }
+}
+
+// .current-img-bar {
+//   margin-left: auto;
+//   margin-right: auto;
+//   width: 100%;
+//   position: absolute;
+//   top: 200px;
+// }
