+++ app/controllers/charges_controller.rb
@@ -5,7 +5,7 @@ class ChargesController < ApplicationController
   def create
     @game = Game.find(params[:game_id])
     @user = User.find(params[:user_id])
-
+    
     item_params = {
       item_id: @game.id,
       buyer_id: @user.id,
@@ -14,8 +14,6 @@ class ChargesController < ApplicationController
 
     @item_transaction = ItemTransaction.new(item_params)
     @item_transaction.save
-
-    puts "Created item transaction: #{item_params}"
 
     @amount = (@game.price * 100).to_i
   
@@ -30,15 +28,17 @@ class ChargesController < ApplicationController
       description: 'Rails Stripe customer',
       currency: 'aud',
     })
+
+    puts "Created item transaction: #{item_params}"
   
+    @game.update(:sold => true)
+
     # @user is the buyer
     # @game.user.id is the seller
 
     NotifyMailer.with(buyer: @user, seller: @game.user, transaction: @item_transaction).confirmation_bought_email.deliver_now
     NotifyMailer.with(buyer: @user, seller: @game.user, transaction: @item_transaction).confirmation_sold_email.deliver_now
 
-    @game.update(:sold => true)
-    
   rescue Stripe::CardError => e
     flash[:error] = e.message
     redirect_to show_game_path(@game.id)
