+++ app/controllers/games_controller.rb
@@ -1,6 +1,7 @@
 class GamesController < ApplicationController
   before_action :set_game, only: [:edit, :update, :show, :destroy]
-  helper_method :find_username_by_id, :get_supported_platforms, :convert_item_condition
+  # helper_method :find_username_by_id, :get_supported_platforms, :convert_item_condition
+  include GamesHelper
   
   # CRUD
 
@@ -42,6 +43,7 @@ class GamesController < ApplicationController
       render 'edit'
       return
     end 
+
     if @game.user != current_user 
       flash[:danger] = "Unauthorized Access"
       redirect_to root_path
@@ -80,45 +82,9 @@ class GamesController < ApplicationController
 
   # Helper Method's
 
-  def find_username_by_id(user_id)
-    
-    if User.exists?(id: user_id)
-      return User.find(user_id).username
-    else
-      # Logic to email support about this issue
-      return "User is missing"
-    end
-  end
-
-  def get_supported_platforms
-    @platforms = ['PlayStation 4', 'Xbox One', 'Nintendo Switch', 'PC']
-  end
-
-  def convert_item_condition(condition_int)
-    case condition_int
-    when 4
-      "Brand New"
-    when 3
-      "Like New"
-    when 2
-      "Very Good"
-    when 1
-      "Good"
-    when 0
-      "Acceptable"
-    else
-      # Logic to email support about this issue
-      "Error: #{condition_int}"
-    end
-  end
-
   # Private params
 
   private
-    def set_game
-      @game = Game.find(params[:id])
-    end
-
     def game_params
       params.permit(:title, :genre, :price, :platform, :condition, :sold, :note, :rating, :img_id, images: [])
     end
