+++ app/views/games/show.html.erb
@@ -1,17 +1,9 @@
 <%= render 'shared/errors', obj: @game %>
-<% if @game.images.attached? %>
-  <div class="listing-images">
-    <% @game.images.each do |file| %>
-      <%= image_tag file, class: 'image-size' %>
-    <% end %>
-  </div>
-<% else %>
-  <p>No images for this game</p>
-<% end %>
+<%= render 'shared/show_images' %>
 
 <h1>Title: <%= @game.title %></h1>
 <p>Genre: <%= @game.genre %></p>
-<p>Price: $<%= number_with_delimiter(sprintf("%.2f", @game.price.round(2))) %></p>
+<p>Price: <%= number_to_currency(@game.price, :unit => "$") %></p>
 <p>Platform: <%= @game.platform %></p>
 <p>Condition: <%= convert_item_condition(@game.condition) %></p>
 <p>Seller Notes: <%= @game.note.to_s %></p>
