+++ app/views/layouts/application.html.erb
@@ -8,9 +8,11 @@
 
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
-
+    
     <%= stylesheet_link_tag    'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= javascript_include_tag 'application', 'data-turbolinks-track': 'reload' %>
+    <%= javascript_include_tag "https://ajax.googleapis.com/ajax/libs/jquery/3.4.0/jquery.min.js", "application" %>
+    <%= javascript_include_tag "https://maxcdn.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js", "application" %>
   </head>
 
   <body>
