+++ app/views/shared/_listing_images.html.erb
@@ -1,0 +1,26 @@
+<div id="game-image-list-<%= game.id %>" class="carousel slide" data-ride="carousel" data-interval="false">
+
+  <!-- Indicators -->
+  <ul class="carousel-indicators current-img-bar">
+    <li data-target="#game-image-list-<%= game.id %>" data-slide-to="0" class="active"></li>
+
+    <% for img_index in 1...game.images.count do %>
+      <li data-target="#game-image-list-<%= game.id %>" data-slide-to="<%= img_index %>"></li>
+    <% end %>
+  </ul>
+
+  <!-- The slideshow -->
+  <div class="carousel-inner">
+    <div class="slideshow-container">
+      <div class="carousel-item active">
+        <%= image_tag game.images[0].variant(resize: '200>x200>'), class: 'image-size' %>
+      </div>
+
+      <% for img_index in 1...game.images.count do %>
+        <div class="carousel-item">
+          <%= image_tag game.images[img_index].variant(resize: '200>x200>'), class: 'image-size' %>
+        </div>
+      <% end %>
+    </div>
+  </div>
+</div>
