+++ app/views/shared/_stripe_form.html.erb
@@ -10,20 +10,15 @@
           </div>
         <% end %>
         
+        <div class="stripe-button-hider">
         <script src="https://checkout.stripe.com/checkout.js" class="stripe-button"
           data-key="<%= Rails.configuration.stripe[:publishable_key] %>"
           data-description="Purchase Game"
-          data-label="Purchase for $<%= number_with_delimiter(sprintf("%.2f", @game.price.round(2))) %>"
+          data-label="Purchase for <%= number_to_currency(@game.price, :unit => "$") %>"
           data-amount="<%= (@game.price * 100) %>"
           data-locale="auto">
         </script>
-        <script>
-          document.getElementsByClassName("stripe-button-el")[0].style.display = 'none';
-        </script>
-
-      <button type="submit" class="btn btn-xs btn-success custom-button-width">
-      Purchase $<%= number_with_delimiter(sprintf("%.2f", @game.price.round(2))) %> AUD
-      </button>
+      </div>
     <% end %>
   <% end %>
 <% else %>
