+++ app/views/users/transactions.html.erb
@@ -5,7 +5,7 @@
 <% @items_purchased.each do |purchased| %>
   <p>Transaction ID: <%= purchased.id %></p>
   <p>Game Title: <%= get_game(purchased.item_id).title %></p>
-  <p>Sold Price: $<%= get_game(purchased.item_id).price %></p>
+  <p>Sold Price: <%= number_to_currency(get_game(purchased.item_id).price, :unit => "$") %></p>
   <p>Seller: <%= get_username(purchased.seller_id) %></p>
   <p>Purchase at: <%= purchased.created_at %></p>
   <hr>
@@ -16,7 +16,7 @@
 <% @items_sold.each do |sold| %>
   <p>Transaction ID: <%= sold.id %></p>
   <p>Game Title: <%= get_game(sold.item_id).title %></p>
-  <p>Sold Price: $<%= get_game(sold.item_id).price %></p>
+  <p>Sold Price: <%= number_to_currency(get_game(sold.item_id).price, :unit => "$") %></p>
   <p>Buyer: <%= get_username(sold.buyer_id) %></p>
   <p>Sold at: <%= sold.created_at %></p>
   <hr>
