+++ db/migrate/20190505021300_change_game_price_to_integer.rb
@@ -1,0 +1,5 @@
+class ChangeGamePriceToInteger < ActiveRecord::Migration[5.2]
+  def change
+    change_column :games, :price, :integer
+  end
+end
