+++ db/migrate/20190505021708_change_game_price_to_integer_two.rb
@@ -1,0 +1,5 @@
+class ChangeGamePriceToIntegerTwo < ActiveRecord::Migration[5.2]
+  def change
+    change_column :games, :price, :decimal, :precision => 8, :scale => 2
+  end
+end
